Save global color to KRA file
Summary:
This is a patch on top of my custom colors patch.

The shared/global assistant color for painting assistants is stored outside of the assistant data model. Because of this I had to add a new property to save to in the KRA saver and loader.

I also had to fix up a few initialization areas in the UI so the loaded values didn't get wiped out.

I also removed a couple "legacy" comments in the KRA saver that got me confused. boud confirmed that those notes were in the calligra days

Test Plan:
Have multiple assistants. A couple with a shared color and one with a custom color.

Saved and loaded the file and made sure it kept the same value.

Reviewers: #krita, dkazakov

Subscribers: dkazakov

Tags: #krita

Differential Revision: https://phabricator.kde.org/D13437