Ten Brushes: Allow users to configure whether to enable the "Preset History" feature
Summary: Ten Brushes in Krita 4.0.4 offers the feature of "Preset History." If one press the shortcut for a preset for the second time, it switches to the previously selected preset. However, some users may be confused with this. This patch adds a check box in the Ten Brushes settings window, and allow users to enable or disable this feature. By default, Preset History will still be turned on.

Test Plan:
[0] Click on "Tools" menu -> "Scripting" -> "Ten Brushes"
Expected: the check box is checked by default.

[1] Open any document in Krita.

[2] Select any brush preset using a shortcut.
Expected: the brush is selected.

[3] Press that shortcut again.
Expected: the previous brush is selected.

[4] Click on "Tools" menu -> "Scripting" -> "Ten Brushes."

[5] Uncheck the check box, and press "OK."
Expected: No error happens, and the dialog closes.

[6] Selevct any other brush preset using a shortcut.
Expected: The brush is selected.

[7] Press that shortcut again.
Expected: The brush is still selected. It does not switch to the previous preset.

[8] Restart Krita, and click on "Tools" menu -> "Scripting" -> "Ten Brushes."
Expected: The settings remain.

[9] Repeat [6] and [7].
Expected: The same as in [6] and [7], respectively.

[10] The testing is done.

Reviewers: #krita, dkazakov

Reviewed By: #krita, dkazakov

Subscribers: dkazakov

Tags: #krita

Differential Revision: https://phabricator.kde.org/D13697