Gamut masking for the artistic color selector
Summary:
Implementation of bug [[ https://bugs.kde.org/show_bug.cgi?id=391210 | 391210 ]].

There is a feature that was not in the design document: the user can set mask strictness in the selector settings to either enforcing (only swatches inside the mask are available for the user to choose) or permissive, "just show the shapes", which only paints the shapes on top of the wheel but does not restrict color picking in any way.

Contrary to the design document, there is no stand-alone mask editor. Instead, when user clicks the edit button, a template opens as a regular view which can be modified, previewed and saved.

**Bugs**
  - ArtColorSel: When initializing the selector for the first time (empty configuration), the wheel segment values are not right

**Smaller tweaks**
  - icon for gamut masking
  - GamutMask: better template design
      - group layer, gray l. with vector on top set to erase
      - set explicit stroke (black) and background (solid transparent) for shapes on preview/save; set fill to white on load to editor, so the erase blend mode works

**Pre-launch**
  - Create user documentation
  - Create pre-made masks, some for basic color harmonies and something fancy to showcase the possibilities

Reviewers: #krita, dkazakov

Subscribers: dkazakov, rempt

Tags: #krita

Differential Revision: https://phabricator.kde.org/D13525