Make it possible for tests to access resources
If a test needs access to the resources in the install folder, it
needs to use KISTEST_MAIN instead of QTEST_MAIN. This will set an
env variable, EXTRA_RESOURCE_DIRS, where KoREsourcePaths will then
also look.

Also remove the setReady/isReady methods for KoResourcePaths: we
don't need that. We can add resource dirs in as many places as
we want.

CCMAIL:kimageshop@kde.org