[bugfix] Connect KisPaintOpBox to the resource manager in constructor
That fixes initialization of the toolbox sliders with correct
values of the lastly used preset. The problem was caused by
the fact that this preset was initialized in cold initialization
of slotInputDeviceChanged(), but the paintop box was not
connected to the resource manager at that time. Calling
restorePreset() later didn't help because the resource manager
already had this preset installed, therefore it didn't emit
any signals again.