Changing the palette docker now sets the document to be modified
But the code is super convoluted... Must be refactored

Currently, KisDocument decides if it's modified after saving by looking at its undostack.
If the undo stack emits sigCleanChanged and tells that it was not clean but is now clean,
the document is unmodified.
This makes sense, as the document might have changed during saving, and that would leave a
trace in the undostack.
But I don't want changing the palette docker leaving a trace in the undostack. The result is
that when saving the document with only modification in the palette, the document won't set
itself back to unmodified.

Now I manually force the document to set itself to unmodified. This is not a good way, though.

T8646