Fix conversion of Pixel Selection into Vector Selection
The patch also fixed a lot of minor issues related to the new
selection overlay functionality when working with vector selections.
Ideally, we should port our KisShapeSelection object to the
full-featured KisShapeLayerCanvas, so that updates would be
done correctly and in real-time. Right now the projection updates
are connected to KisUpdateSelectionJob that is issues once in 300
milliseconds.

BUG:396470