Fix the KisUpdaterContextTest
Because the KisUpdaterContext member variable of KisUpdateScheduler
is no longer a pointer, it's not possible to replace the existing
KisUpdaterContext with KisTestableUpdaterContext; instead merge
the test functionality directly into KisUpdaterContext and use
that from the kis_update_scheduler_test test.

The other unittests do use the KisTestableUpdaterContext because
otherwise things start going off for real, and that breaks those
tests.