Fix assert when undoing Convert Vector to Selection action
The problem is that updates of the selections can be executed only
asynchronously via KisImage's scheduler. It means that we cannot
update a selection while it is not attached to any selection mask
that is a part of image's hierarchy.

Therefore now, if we get a selection update request while the
selection is not part of the layer hierarchy, we just save this
request and mark it as "stalled". As soon as a selection gets
attached to the image, the updtate is issued.