Resolve conflicts between create and move selection actions
We have a really weird design of the selection tools.
Basically, it happens because we just wrap Ko- tools
into some wrappers and inherit from them. Ideally, all
the tools should use begin/continue/endPrimaryAction()
and/or implement KoInteractionTool pattern.

NOTE: Select Path tool does not support moving of the
selection, because it is not possible to implement it
without a complete refactoring of KoCreatePathTool.