Add a workaround for tablets not reporting tablet events in hover mode
The bug is caused by the fact that we postpone tablet events
by one event (presumably to make it sync with mouse events
stream). But some non-wacom tablets do not report tablet
move events when the stylus is hovering. It means that the
following tablet press event will be positioned incorrectly.

This event postponing has come from Qt, and the commit message
says it was needed for "relative" mode, which we don't support.

I just disabled this postponing, let's check whether it helps
people.

BUG:363284