Use pyqt snapshots
The official release cannot be built:

" There is a bug (fixed in the current PyQt5 snapshot) that happens if you
are building PyQt5 for the very first time. The workaround is to create
an empty __init__.py file in the same directory containing sip.pyd."

And we're always building for the very first time, of course.