Avoid blending bug on first color pick.
In a newly opened Krita instance, the starting foreground color is
always in a 16-bit RGBA16 format, regardless of the color space of newly
created/opened Krita document. As such, using the color picker's
blending feature (which mixes a newly picked color with the user's
previous color) immediately after the program starts results in an
incorrect or undefined output color.

This bug mitigates the problem and fixes the bug request by simply
restricting to "pure" (without radius or blending) picking on the first
use.

BUG: 394399 "First time picking a color..."