Adjust selected curve point more accurately by using spin boxes
Summary:
In order to adjust selected curve point accurately, spin boxes are added
on x and y domain. Selected curve point can be re-positioned by editing
these spin boxes.

For non-numeric values and where the maximum value is 1.0, percentage
is used; for time values in seconds, milliseconds is used. In
order to emphasize this, suffixes are added to spin boxes.

BUG: 396528

{F6309671} {F6309672}

Reviewers: #krita, #vdg, scottpetrovic, rempt

Reviewed By: #krita, scottpetrovic, rempt

Subscribers: rempt, langkamp, scottpetrovic

Tags: #krita

Differential Revision: https://phabricator.kde.org/D16008