remove a 1-iteration loop in  ui/KisAsyncAnimationFramesSavingRenderer
Summary: While trying to read the animation code to understand how rendering is done, I found this 1-iteration loop. At first I thought that the `for` loop was iterating over all the frames (rendering to files synchronously). I think the code is easier to read with the false loop removed.

Test Plan: (To be honest, I haven't built the resulting code. I was reading the code from a git clone, and this is a minor code change, but I haven't managed to setup a proper build environment for Krita yet. There are animation-rendering tests in place, so I'm confident that Continuous Integration would catch any potential issue here.)

Reviewers: dkazakov, #krita

Reviewed By: dkazakov, #krita

Tags: #krita

Differential Revision: https://phabricator.kde.org/D16201