add a createActions method to KoToolFactoryBase
Actions created by a tool instance only exist if there's an open
document, which means they cannot be configured or placed in
toolbars. We should only have per-window actions, not per-action
actions.

Creating the actions in the toolfactory instead of the tool when
the window is created fixes that: now the tools will retrieve
the actions from the global action collection on activations
and de-activation, like the freehand tool already did for increase
and decrease brush size.

Note: this commit breaks the following tools because it's not
worth porting hidden tools in any case:

* artistic text tool
* odf text tool
* connection tool

BUG:365222

CMAIL:kimageshop@kde.org