BUG:365222 Fix activating actions if a shortcut scheme is loaded
Loading custom shortcuts sets the shortcuts for actions that are
stripped of their default shortcuts to "", an empty string. This
means that the shortcut is not empty, so the logic that checks
for duplicate shortcuts that need to be disabled when a tool
overrides the global shortcuts thinks that all actions with a ""
shortcut need to be disabled.