Work around QTBUG-72488
Qt 5.12 now deletes a void pointer member variable of QTabletEvent.
QTableEvent doesn't implement a copy constructor for QTabletEvent,
but Krita does copy events when compressing move events. Both
the original and the copy try to delete the pointer, and poof.

This is strictly a workaround, only valid for Qt 5.12 and should
be removed if this bug is fixed in Qt 5.12.1 or later.

BUG:401988,401988