Insert new layer above selected group layer
Summary: About a month ago or something I said in the irc channel that
I'd t ake a look at inserting the new layers above the group layer if th
e group layer is selected as opposed to inserting it in the group. The
rationale is that you can already easily insert the layer in the group
by selecting any layer inside that group, but if you want to insert the
layer just above the group you're forced to do it in two steps: first
insert the layer inside the group, then move the layer above the
group (eg. with ctrl+pgup).  It's not a big deal, but I know at least
one person that mentioned this before (don't remember if it was in irc
or opened a bug).

In any case it's a very tiny change that I finally got around to doing it.

The patch by Razvan Radulescu (https://phabricator.kde.org/p/razvanr/)

BUG:399103
Reviewers: rempt, scottpetrovic, #krita
Reviewed By: rempt, #krita
Tags: #krita
Differential Revision: https://phabricator.kde.org/D15523