krita.desktop - added StartupWMClass
Without this patch, when the icon is added to the launcher, in Ubuntu (gnome), and it is clicked, then a new icon is added dynamically to the pane, with the dot on its left showing the new icon as an active application. With this patch, this is fixed, and now when I click on the Krita icon in the launcher, I just get a dot on its left, showing it as an active application.