Advanced Color Selector: option to hide color selector
Summary:
I have resurrected the original patch by Ranveer Aggarwal, which enables the user to hide the
color selector part of the docker.

There was one issue with the original: the docker settings button is inside the color selector, therefore when you hide
the selector, the only way to configure the docker is via the main menu, "Settings" -> "Configure Krita". To solve that I have added
a fallback settings button to the docker widget, which appears when the color selector part is hidden.

BUG: 337328

Reviewers: #krita, rempt

Reviewed By: #krita, rempt

Subscribers: rempt

Tags: #krita

Differential Revision: https://phabricator.kde.org/D17893