Do not create the move tool option widget twice
Flake first activates a tool, then asks it to create its option
widget(s). However, the move tool needs to have the option widget
created when it's activated, so create the option widget in the
constructor of the move tool.

This also incidentally fixes the problems with actions only being
executed after changing the current node having used the move tool.

BUG:403048