Changed The Properties for Generator (Fill) Layers to be Non-Modal
This change moves toward consistency with other properties windows
allowing you to move the properties window of a fill layer. The new window
can be called either as modal or non modal depending on whether or not it has
existing layer data passed into it. This means that creating a new fill layer
uses a modal window while editing allows you to move the window to the side to compare
and contrast values.