Use QuaZip instead of KArchive for ZIP64 support
This makes it possible to save and load .kra files that are larger
than 4GiB -- depending on available memory, of course. There is
an option in the settings dialog to enable this; it's off by default
because older versions of Krita cannot read Zip64 files.

Note that everything that uses zip files now uses quazip, the
karchive dependency is gone.

CCMAIL:kimageshop@kde.org