Add direction choice to the imagesplit plugin
One use-case for the imagesplit is modifying spritesheet found externally
(for example on the internet): splitting them then re-importing them in
the animation timeline lets you use the dedicated animation tools, like
the onionskin, and easily preview the animation with your modifications.

But most spritesheet found on the internet are horizontal, like so:

1, 2, 3
4, 5, 6

while a few are vertical, like so:

1, 3, 5
2, 4, 6

Adding a button to choose which is more appropriate lets the imagesplit
be used in both cases and still name the individual images correctly.

I put the default at horizontal, while it was at vertical before,
because it seems more widely used.

Differential revision: https://phabricator.kde.org/D18442