Temporarily push D18467
This patch changes the way how opacity and flow interact. This can mean
that a lot of presets work way differently from before. Therefore artists
need to test so we can get some feedback on how the patch feels.

For more information, and to give feedback, check https://phabricator.kde.org/T8576

This is so that artists can test it via the nightly builds.

We'll try to revert it on Monday 28th of January.

Patch by William Brown.

Note: We still need to do a proper code review!

CCMAIL:kimageshop@kde.org
Ref T8576