Imeplemented automatic selection of a renderer and/or surface format
This patch generalizes the thing previously implemented in kis_opengl_win
by Alvin Wong long ago.

The algorithm is based on two finctions: KisOpenGL::selectSurfaceFormat()
and KisOpenGLModeProber::probeFormat(). The former one generates the full
list of all possible renderer/format combinations, then sorts it according
to user's preferences and HDR necessity, then probes every resulting format
with the latter function. As soon as it finds a format that fits the
requirements, it makes this format default for the root surface of the
application.