Added Real-Time Preview when Changing Generator Layer Properties.
Colors and patterns will update when changing the values within the
properties window of a generator layer. The changes will only be registered
with the undo command system when accepting the change. Canceling your changes
will revert the layer back to the state before the properties panel was opened.

This should make changing the fill layer color a more artistic process,
letting you visually see the effects before applying the change.