BUG:399796 Disables use of Always Use Template in the New File window
When new custom template is created, Always Use Template and Full Template Path
are both set to the new template. When a template is being used, only
Full Template Path is set to the last used template. When highlighting
the template as default to use in New File window, Krita first uses
Always Use Template, then Full Template Path, then just first template
in the alphabetical order. To make sure the last used template is always the one
that Krita highlights, Always Use Template needs not to be considered.
This commit removes the mention of Always Use Template
in the New File window's code.

Differential Revision: https://phabricator.kde.org/D18812