Make all aspects of autosaving and backup files configurable
Autosaves of unnamed files can now be made visible (the dot
is omitted).

The number of backups kept can be set, the default is 1 again.
The location of backups can be chosen: same folder as the file,
$HOME or the tempdir.
The suffix used for the backup file can be chosen. There is
a validator to weed out potentially hazardous characters.

CCMAIL:kimageshop@kde.org
BUG:404395
Differential Revision: https://phabricator.kde.org/D19115