Restore icon for tag-dropdown in pop-up palette.
While I am at it, I've replaced it with the breeze tag icon, as the
settings icon did not much make sense since it was replaced from a wrench
to it's current mini representation of an image window.

Also chose the 16x16 version of this as it's easier to read.