Fix problems when loading Krita 4.1 files saved in non-utf-8 locales
Krita 4.1 has a bug where when the locale is not utf-8, the entries
in the zip file entrylist are in the local locale, not utf-8. The
image name is encoded in utf-8 in the maindoc.xml file, but the
layers are, because of ancient history when more than one image
could be stored in a kra file, in a subfolder with the image name.

So, if the image name in utf-8 is different from the image name
in the directory, quazip gets confused because it actually handles
encoding correctly.

Ark also cannot load kra files that are borked in this with.

This hack substitutes the encoded name for the utf-8 name in the
quazip store implementation, and now the files can load again.

BUG:405475