Fix selecting presets (more or less)
Clicking in empty space will still give a crash.

Note: reloading the preset if dirtypresets is unchecked is broken
now. There's a warning about that. I'm not sure whether we even
want that possibility, or would prefer a right-click to restore
a preset to default or something like that in the preset chooser.