Fix a race condition in KisLocklessStack
When recycling the objects after pop(), a race condition could happen.
After the thread has successfully passed 'if (m_deleteBlockers == 1)'
check, it could go to sleep. During its sleep some other threads could
release more objects into m_freeNodes, which were still accessed by
someone. It could lead to both: invalid access or even ABA problem
(the latter, theoretically, could cause bug 405663).

The patch fixes the problem by forcing the releasing thread to check
if it is still the only owner of the freeNodes list. It atomically
fetches the list of free nodes, and only after that checks if all its
other users are dead. Since the list is stored in a local variable,
it is guaranteed that noone modifies it during the check.

BUG:405663