Fix race condition in the lockfree hash table's garbage collector
This is the same problem that has been present in the lockfree
stack until recently. We should check for m_rawPointerUsers when
we already have a local copy of the free nodes list. Only in this
case it guarantees that noone added elements between the check and
the actual destruction.