Disable assert on empty parent in Colorize Mask
Before this change, Colorize Mask asserted after flattening
or merging with another layer when the image was smaller than 1024px
wide and long. After this commit, it shouldn't happen.

It was caused by setting the parent dirty just before the parent gets
deleted. The request was linked to refreshing area outside the image
boundaries; since Colorize Mask didn't and doesn't work outside of the
image boundaries, it isn't necessary.

NOTE: In case of providing support for Colorize Mask working outside of
the canvas boundaries, the decision in this commit might need
reevaluation (in case of a need to update the mask there).

BUG:402494