BUG:406032 Fix crash caused by unsupported composition mode in QT 5.12
Qt 5.12 doesn't support QPainter::RasterOp_SourceXorDestination anymore.

Trying to paint with a painter that uses this while opengl is enabled
will lead to a crash. The only relevant place this enum value is still used
is in kis_tool.cc, where it is only painted on the regular non-accelerated canvas.

Anyhow, told the smartpatch to use this function instead of doing something itself.

The whole thing is a little bizar.