Fix wrong DPI when moving window across screens
This fixes a bug causing the wrong DPI being used when moving the window
from a screen to another screen to the left. Using the top-left position
of the window to determine screen changes doesn't work well since the
device pixel ratio of a window does not change until the window is
roughly halfway across.

Maniphest Tasks: T2299
Differential Revision: https://phabricator.kde.org/D20171