Make Windows build create thin archives when linking
This makes the Windows build create thin archives when linking the
binaries to reduce the amount of redundant writes to disk when building.

With a normal clean build, the size of the build directory has reduced
by 2GB.

When I tested building on a RAM-based filesystem, the clean build time
is reduced from 52 minutes to 50. The result could be more significant
on SSD/HDD.

(Above data does not include building the deps.)

Note: If you encountered build errors, you may need to do a clean
rebuild. This is only for Windows.

CCMAIL: kimageshop@kde.org