Fix page orientation switch in new image dialog
Summary:
KisDoubleParseSpinBox does not update it's state while being edited (to
allow for the math formulas to work correctly). If the cursor is in the
number field while the user is changing page orientation, the focused
spinbox's value does not change. To work around that, this patch clears
focus on the width and height spinboxes before switching page
orientation.

BUG: 374493

Reviewers: #krita, rempt

Reviewed By: #krita, rempt

Tags: #krita

Differential Revision: https://phabricator.kde.org/D20300