Fix fullscreen mode on Windows with OpenGL
A previous attempt was made in D1977 [1] but it didn't work properly.
The patch `0003-Hack-for-fullscreen-workaround.patch` only partly hid
the issue and the issue has resurfaced since this patch was removed.

This commit actually enables Qt's workaround with an additional patch
from Qt 5.13 [2]. The workaround adds a 1px border around the window. An
additional hack is added to hide three of the borders, though we can't
hide all four borders since the bug will return. This currently leaves
the bottom border visible as a compromise.

This workaround is not active when ANGLE is being used. If anyone is
annoyed by the 1px border, they may consider switching to ANGLE.

[1]: https://phabricator.kde.org/D1977
[2]: https://codereview.qt-project.org/246155

Differential Revision: https://phabricator.kde.org/D20657