Fix asserts in KisSuspendProjectionUpdatesStrokeStrategy
It looks like RESUME stroke strategy works differently from
other strategies. It can be suspended even when there are no
internal jobs left. It happens because the strokes queue uses
RESUME strategies as a marker for a position, where to place
LodN strokes. It will add new strokes before it even if the
resume strategy is almost done :)

BUG:404552