Add a hidden config option to set the default exr F16/F32 profile
Exr files don't have profiles, and profiles are totally irrelevant
for exr/hdr workflow: the lut docker is what is used in these cases.
But since we need to set a profile on a krita image, let's make it
configurable at least.

Note: there's no gui for this because a) it's an advanced option
and b) we're in string freeze.

CCMAIL:kimageshop@kde.org
BUG:406939