Remove the histogram dialog
The histogram dialog shows incorrect histograms, compared to the
histogram docker. The histogram docker is also prettier and more
clear.

We do lose the ability to show histograms in diffrent colorspaces
than the layer's colorspace, but that seems to be even more broken.

We also cannot scale histograms for floating point colorspaces
anymore, but that also seemed to be broken.