Implement a projection store for masks (for indirect painting)
We have implemented correct painting on masks, but we have never
implemented a projection for them. Indeed, for normal rendering
process they don't need it. But when a mask is selected in isolated
mode, it still needs a projection, so that the use could see the result
of the painting.

BUG:403576