Add a dialog to change the source of clone layers
This commit adds an action named `Set Copy From' to the context menu
of the Layers Docker, if one or more clone layers are selected. The
action pops up a dialog that can change the `Copy From' property of
selected clone layers.

The dialog previews the changes to these layers.
If there is any change to the document when the dialog is open, the
changes in the dialog will be applied and the dialog closed.