Complete handling of all fatalX errors in xcf lib
Before this commit some fatalX calls didn't have return
just after the call, because the library assumed that xcf
will die anyway. Since we want to return an error instead
of killing xcf (and Krita in the process), we need to assure
that all fatalX calls are followed by return with some indication
of an error. This commit does exactly that.

There are still some functions that returns int (error/ok code)
but the return value is not checked by callers.