Feature: 'Reference Tool' now allows loading images directly from clipboard.
More work can / should be done here. It would be nice if URLs to image
files could be loaded as if it were another file. Same with copying a
file from a local directory. This could perhaps be represented as a
link? This would require a unique 'try loading image' loop that can
gracefully fail and inform the user without locking up the program.

Testing: Try this feature on Windows and Mac OSX. This only works when
the image itself is copied into the clipboard, not the url or reference
to the image.