Disable checking the unique ID for the stylus
The Qt documentation says "As of Qt 4.2, the unique ID is the same
regardless of the orientation of the pen. Earlier versions would report
a different value when using the eraser-end versus the pen-end of the
stylus on some OS's." (https://doc.qt.io/qt-5/qtabletevent.html#uniqueId)

As the log in CCBUG:407659 shows, this is not true.