Fix crash in move tool after closing the document
We shouldn't use lambda functions in signal connection if we need
to control lifetime of the connection. In this very case, the connection
was still active after destruciton of KisToolMove and lambda still stored
a pointer to destroyed 'this'.

BUG:407228
BACKPORT:krita/4.2
(cherry picked from commit 4622b00348ce7973f53245930506aef5099c9850)