Change edit guides cmd to post-exec & auto-upload to doc
KisChangeGuidesCommand is now post-exec, which means the first redo
will be ignored. The command will not be created during dragging.
Instead, only when the mouse/pen/finger is released will the command
be created. The commands will not be merged.

Any set* function of KisGuidesManager will upload
the config to KisDocument.