Make KisSafeNodeProjectionStore actually reuse the projection
Without any image present, the store will just drop the
projection without any reusing. It may cause small slowdowns
in the beginning of the stroke and (theoretically) memory
fragmentation on Windows.

I'm not sure if it actually the reason of the 408133, so we
need a confirmation from the reporter first.

CCBUG:408133